server finished

+++ app/redirect/models/redirect_info.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, event, text
+from sqlalchemy import Column, Integer, BigInteger, String, event, text
 from sqlalchemy.sql import func
 import sqlalchemy
 
@@ -8,7 +8,7 @@ from core.database import Base
 class RedirectInfo(Base):
     __tablename__ = "redirect_info"
 
-    server_id = Column(Integer, primary_key=True, index=True)
+    server_id = Column(BigInteger, primary_key=True, index=True)
     server_link = Column(String, nullable=False, unique=True, index=True)
     domen_link = Column(String, nullable=False, unique=True, index=True)
     last_use = Column(
